Add CityHistoryBox, wip: CityCompareModal
